update and getById service

+++ src/main/java/dk/amir/dev/repository/ContactRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ContactRepository extends JpaRepository<Contact, Integer> {
+public interface ContactRepository extends JpaRepository<Contact, Long> {
     List<Contact> findByType(ContactType type);
 }
